# swagger.yaml
openapi: 3.0.0
paths:
...
  /signin:
    post:
      operationId: AuthController_signin
      summary: Войти в систему
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Успешный вход в систему
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    description: Идентификатор
                  username:
                    type: string
                    description: Имя пользователя
                  phone:
                    type: string
                    description: Номер телефона
                  avatar:
                    type: string
                    description: URL аватара
                  balance:
                    type: number
                    description: Баланс
                  accounts:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          description: Идентификатор аккаунта
                        type:
                          type: string
                          description: Тип аккаунта
                        role:
                          type: string
                          description: Роль аккаунта
                        credentials:
                          type: object
                          properties:
                            email:
                              type: string
                              description: Email
                            accessToken:
                              type: string
                              description: AccessToken
                            refreshToken:
                              type: string
                              description: RefreshToken
                        profile:
                          type: string
                          description: Идентификатор профиля
                    description: Список аккаунтов
        '401':
          description: Неверное имя пользователя или пароль
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Неверное имя пользователя или пароль
                    description: Сообщение об ошибке
                  error:
                    type: string
                    example: Unauthorized
                    description: Тип ошибки
                  statusCode:
                    type: number
                    example: '401'
                    description: HTTP-статус код
      tags: &ref_0
        - auth
  /signup:
    post:
      operationId: AuthController_signup
      summary: Регистрация
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                username:
                  type: string
      responses:
        '201':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    description: Имя пользователя
                  phone:
                    type: string
                    description: Номер телефона
                  avatar:
                    type: string
                    description: URL аватара
                  balance:
                    type: number
                    description: Баланс
                  accounts:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          description: Тип аккаунта
                        role:
                          type: string
                          description: Роль
                        credentials:
                          type: object
                          properties:
                            email:
                              type: string
                              description: Email
                            accessToken:
                              type: string
                              description: AccessToken
                            refreshToken:
                              type: string
                              description: RefreshToken
                        profile:
                          type: string
                          description: 'Идентификатор профиля '
                        _id:
                          type: string
                          description: Идентификатор аккаунта
                    description: Список аккаунтов
                  _id:
                    type: string
                    description: Идентификатор профиля
        '400':
          description: Некорректные данные
        '409':
          description: Аккаунт уже существует
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Аккаунт уже существует
                    description: Сообщение об ошибке
                  error:
                    type: string
                    example: Conflict
                    description: Тип ошибки
                  statusCode:
                    type: number
                    example: '409'
                    description: HTTP-статус код
      tags: *ref_0
  /refresh-token:
    post:
      operationId: AuthController_refreshToken
      summary: Обновить токен
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '201':
          description: accessToken успешно обновлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    description: accessToken по умолчанию действует 1 день
                  refreshToken:
                    type: string
                    description: refreshToken по умолчанию действует 7 дней
        '401':
          description: Невалидный refreshToken
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Невалидный refreshToken
                    description: Сообщение об ошибке
                  error:
                    type: string
                    example: Unauthorized
                    description: Тип ошибки
                  statusCode:
                    type: number
                    example: '401'
                    description: HTTP-статус код
      tags: *ref_0
...
  /yandex/success:
    get:
      operationId: AuthController_yandexSuccess
      parameters:
        - name: token
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      tags: *ref_0
...
  /reset-password:
    post:
      operationId: AuthController_resetPassword
      summary: Сброс пароля
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: Email пользователя
      responses:
        '201':
          description: Ссылка для сброса пароля отправлена на указанный Email
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: >-
                      Ссылка на сброс пароля отправлена на ваш email:
                      youremail@domain.ru
                    description: Email пользователя
        '400':
          description: Некорректные данные
        '404':
          description: Пользователь с указанным Email не найден
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Пользователь с указанным Email не найден
                    description: Сообщение об ошибке
                  error:
                    type: string
                    example: Not Found
                    description: Тип ошибки
                  statusCode:
                    type: number
                    example: '404'
                    description: HTTP-статус код
      tags: *ref_0
...
info:
  title: API BotKits
  description: Ручки для команды frontend
  version: '1.0'
  contact: {}
...
servers: []
...